fix: delete detailView 컴포넌트, 페이지로 이동
